added client and controller to get all messages from db
