+++ .gitignore
@@ -2,5 +2,3 @@ node_modules
 .env
 test-miguel.md
 data
-service/utils/messageQueue.js
-service/utils/MessageService.js
