+++ service/clients/findAllMessages.js
@@ -1,0 +1,7 @@
+const Message = require('../models/Message');
+
+const findAllMessages = () => {
+    return Message.find()
+}
+
+module.exports = findAllMessages;
